[GEOS-10269] Overriding Json Object during features templating merge operation
