+++ src/community/features-templating/features-templating-core/src/main/java/org/geoserver/featurestemplating/builders/TemplateBuilderMaker.java
@@ -12,6 +12,7 @@ import org.geoserver.featurestemplating.builders.flat.FlatDynamicBuilder;
 import org.geoserver.featurestemplating.builders.flat.FlatIteratingBuilder;
 import org.geoserver.featurestemplating.builders.flat.FlatStaticBuilder;
 import org.geoserver.featurestemplating.builders.impl.CompositeBuilder;
+import org.geoserver.featurestemplating.builders.impl.DynamicMergeBuilder;
 import org.geoserver.featurestemplating.builders.impl.DynamicValueBuilder;
 import org.geoserver.featurestemplating.builders.impl.IteratingBuilder;
 import org.geoserver.featurestemplating.builders.impl.RootBuilder;
@@ -52,6 +53,10 @@ public class TemplateBuilderMaker {
 
     private String separator = "_";
 
+    private boolean isDynamicMerge = false;
+
+    private JsonNode baseMergeNode;
+
     public TemplateBuilderMaker() {
         this.encondingHints = new EncodingHints();
         this.vendorOptions = new VendorOptions();
@@ -84,6 +89,18 @@ public class TemplateBuilderMaker {
      */
     public TemplateBuilderMaker jsonNode(JsonNode jsonNode) {
         this.jsonNode = jsonNode;
+        return this;
+    }
+
+    /**
+     * Set the baseMergeNode for the builder to be created. Only DynamicMergeBuilder can have a
+     * baseMergeNode content.
+     *
+     * @param baseMergeNode the dynamic merge content
+     * @return this TemplateBuilderMaker
+     */
+    public TemplateBuilderMaker baseMergeNode(JsonNode baseMergeNode) {
+        this.baseMergeNode = baseMergeNode;
         return this;
     }
 
@@ -260,6 +277,17 @@ public class TemplateBuilderMaker {
         return this;
     }
 
+    /**
+     * Set to true if the builder if its a merge operation that requires dynamic merging.
+     *
+     * @param isDynamicMerge
+     * @return
+     */
+    public TemplateBuilderMaker enableDynamicMerge(boolean isDynamicMerge) {
+        this.isDynamicMerge = isDynamicMerge;
+        return this;
+    }
+
     /** Reset all the attributes of this TemplateBuilderMaker. */
     public void globalReset() {
         localReset();
@@ -281,6 +309,7 @@ public class TemplateBuilderMaker {
         this.jsonNode = null;
         this.rootBuilder = false;
         this.topLevelFeature = false;
+        this.isDynamicMerge = false;
     }
 
     /**
@@ -339,6 +368,15 @@ public class TemplateBuilderMaker {
         return dynamicValueBuilder;
     }
 
+    private DynamicMergeBuilder buildDynamicMergeBuilder() {
+        DynamicMergeBuilder dynamicMergeBuilder =
+                new DynamicMergeBuilder(name, textContent, namespaces, baseMergeNode);
+        if (filter != null) dynamicMergeBuilder.setFilter(filter);
+        if (!encondingHints.isEmpty())
+            dynamicMergeBuilder.getEncodingHints().putAll(encondingHints);
+        return dynamicMergeBuilder;
+    }
+
     private StaticBuilder buildStaticBuilder() {
         StaticBuilder staticBuilder;
         boolean hasJsonNode = jsonNode != null;
@@ -372,6 +410,8 @@ public class TemplateBuilderMaker {
         else if (textContent == null && jsonNode == null) {
             if (isCollection) result = buildIteratingBuilder();
             else result = buildCompositeBuilder();
+        } else if (isDynamicMerge) {
+            result = buildDynamicMergeBuilder();
         } else {
             if (textContent != null && textContent.contains(TemplateReader.EXPRSTART))
                 result = buildDynamicBuilder();
